new3: Q20 GRU type migration - ctfa TA/FA output uint32_t, fusion >>20
Changes:
- ulunas_fp.c: ctfa_ta_module/ctfa_fa_module output uint16_t->uint32_t
- ulunas_fp.c: sigmoid_q20_to_q15->sigmoid_q20_to_q20 in TA/FA
- ulunas_fp.h: declarations updated to uint32_t* y
- ulunas_modules.c: all y_ta/y_fa uint16_t->uint32_t
- ulunas_modules.c: fusion r=16384->524288, >>15->>>20
- test files: same type/shift updates

Known: GRU output still int16_t (Q15) limits TA/FA precision.
Next: change gru_module/bigru_module output int16_t->int32_t (Q20).

+++ new3/x2000_deploy_v2/dump_full_pipeline.c
@@ -254,7 +254,7 @@ int main(int argc, char **argv) {
             }
 
             /* E0.4: cTFA TA */
-            uint16_t y_ta[12];
+            uint32_t y_ta[12];
             ctfa_ta_module(y_ap, 12, 65, E0_CTFA_TA_GRU_NHID,
                            st.tfa_cache_e0,
                            encoder_en_convs_0_ops_4_ta_gru_weight_ih_l0,
@@ -338,7 +338,7 @@ int main(int argc, char **argv) {
             }
 
             /* E0.5: cTFA FA */
-            uint16_t y_fa[65];
+            uint32_t y_fa[65];
             ctfa_fa_module(y_ap, 12, 65, E0_CTFA_FA_GRU_NHID,
                            E0_CTFA_FA_GROUP, E0_CTFA_FA_SEG, E0_CTFA_FA_PAD,
                            encoder_en_convs_0_ops_4_fa_gru_weight_ih_l0,
